requirements.txt
Django>=4.1,<5
django-cors-headers
djangorestframework